Удалит старый suite
