+++ internal/usecases/suite/service_suite.go
@@ -1,12 +1,8 @@
 package serviceSuite
 
 import (
-	"context"
 	"errors"
 	"math/rand"
-	"os"
-	"path/filepath"
-	"runtime"
 	"slices"
 	"time"
 
@@ -14,124 +10,84 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/mock"
 	testifySuite "github.com/stretchr/testify/suite"
-	"github.com/testcontainers/testcontainers-go/modules/postgres"
 
 	"github.com/nice-pea/npchat/internal/common"
 	"github.com/nice-pea/npchat/internal/domain/chatt"
+	mockChatt "github.com/nice-pea/npchat/internal/domain/chatt/mocks"
+	mockSessionn "github.com/nice-pea/npchat/internal/domain/sessionn/mocks"
+	mockUserr "github.com/nice-pea/npchat/internal/domain/userr/mocks"
+
 	"github.com/nice-pea/npchat/internal/domain/sessionn"
 	"github.com/nice-pea/npchat/internal/domain/userr"
-	pgsqlRepository "github.com/nice-pea/npchat/internal/repository/pgsql_repository"
 	"github.com/nice-pea/npchat/internal/usecases/events"
-	"github.com/nice-pea/npchat/internal/usecases/users/oauth"
 	mockOauth "github.com/nice-pea/npchat/internal/usecases/users/oauth/mocks"
 )
 
-type Suite struct {
+type SuiteWithMocks struct {
 	testifySuite.Suite
-	factory       *pgsqlRepository.Factory
-	factoryCloser func()
-	RR            struct {
-		Chats    chatt.Repository
-		Sessions sessionn.Repository
-		Users    userr.Repository
+	RR struct {
+		Chats    *mockChatt.Repository
+		Sessions *mockSessionn.Repository
+		Users    *mockUserr.Repository
 	}
 	Adapters struct {
-		Oauth oauth.Provider
+		Oauth *mockOauth.Provider
 	}
 	MockOauthTokens map[string]userr.OpenAuthToken
 	MockOauthUsers  map[userr.OpenAuthToken]userr.OpenAuthUser
 }
 
-var (
-	pgsqlDSN = os.Getenv("TEST_PGSQL_DSN")
-)
-
-// newPgsqlExternalFactory создает фабрику репозиториев для тестирования, реализованных с помощью подключения к postgres по DSN
-func (suite *Suite) newPgsqlExternalFactory(dsn string) (*pgsqlRepository.Factory, func()) {
-	factory, err := pgsqlRepository.InitFactory(pgsqlRepository.Config{
-		DSN: dsn,
-	})
-	suite.Require().NoError(err)
-
-	return factory, func() { _ = factory.Close() }
-}
-
-// newPgsqlContainerFactory создает фабрику репозиториев для тестирования, реализованных с помощью postgres контейнеров
-func (suite *Suite) newPgsqlContainerFactory() (f *pgsqlRepository.Factory, closer func()) {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	// Поиск скриптов с миграциями
-	_, b, _, _ := runtime.Caller(0)
-	migrationsDir := filepath.Join(filepath.Dir(b), "../../../infra/pgsql/init/*.up.sql")
-	migrations, err := filepath.Glob(migrationsDir)
-	suite.Require().NoError(err)
-	suite.Require().NotZero(migrations)
-
-	postgresContainer, err := postgres.Run(ctx,
-		"postgres:17",
-		postgres.WithInitScripts(migrations...),
-		postgres.WithDatabase("test_npc_db"),
-		postgres.WithUsername("test_npc_user"),
-		postgres.WithPassword("test_npc_password"),
-		postgres.BasicWaitStrategies(),
-	)
-	suite.Require().NoError(err)
-	dsn, err := postgresContainer.ConnectionString(ctx, "sslmode=disable")
-	suite.Require().NoError(err)
-
-	f, err = pgsqlRepository.InitFactory(pgsqlRepository.Config{
-		DSN: dsn,
-	})
-	suite.Require().NoError(err)
-
-	return f, func() {
-		_ = f.Close()
-		_ = postgresContainer.Terminate(context.Background())
-	}
-}
-
 // SetupTest выполняется перед каждым тестом, связанным с suite
-func (suite *Suite) SetupTest() {
-	// Инициализация фабрики репозиториев
-	if pgsqlDSN != "" {
-		suite.factory, suite.factoryCloser = suite.newPgsqlExternalFactory(pgsqlDSN)
-	} else {
-		suite.factory, suite.factoryCloser = suite.newPgsqlContainerFactory()
-	}
+func (suite *SuiteWithMocks) SetupTest() {
+	suite.TearDownSubTest()
+}
 
-	// Инициализация репозиториев
-	suite.RR.Chats = suite.factory.NewChattRepository()
-	suite.RR.Users = suite.factory.NewUserrRepository()
-	suite.RR.Sessions = suite.factory.NewSessionnRepository()
+// SetupAcceptInvitationMocks настраивает моки для успешного принятия приглашения
+func (suite *SuiteWithMocks) SetupAcceptInvitationMocks(invitationID uuid.UUID, chat chatt.Chat) {
+	suite.RR.Chats.EXPECT().List(chatt.Filter{
+		InvitationID: invitationID,
+	}).Return([]chatt.Chat{chat}, nil).Once()
 
-	// Инициализация адаптеров
+	suite.RR.Chats.EXPECT().Upsert(mock.Anything).RunAndReturn(func(updatedChat chatt.Chat) error {
+		suite.Equal(chat.ID, updatedChat.ID)
+		return nil
+	}).Once()
+}
+
+// TearDownSubTest выполняется после каждого подтеста, связанного с suite
+func (suite *SuiteWithMocks) TearDownSubTest() {
+	// пересоздаем моки репозиториев
+	suite.RR.Chats = mockChatt.NewRepository(suite.T())
+	suite.RR.Users = mockUserr.NewRepository(suite.T())
+	suite.RR.Sessions = mockSessionn.NewRepository(suite.T())
 	suite.Adapters.Oauth = mockOauth.NewProvider(suite.T())
-	suite.Adapters.Oauth.(*mockOauth.Provider).
-		On("Name", mock.Anything).Maybe().
-		Return("mock").
-		On("Exchange", mock.Anything).Maybe().
-		Return(func(code string) (userr.OpenAuthToken, error) {
-			token, ok := suite.MockOauthTokens[code]
-			if !ok {
-				return userr.OpenAuthToken{}, errors.New("token not found")
-			}
-			return token, nil
-		}).
-		On("User", mock.Anything).Maybe().
-		Return(func(token userr.OpenAuthToken) (userr.OpenAuthUser, error) {
-			user, ok := suite.MockOauthUsers[token]
-			if !ok {
-				return userr.OpenAuthUser{}, errors.New("user not found")
-			}
-			return user, nil
-		}).
-		On("AuthorizationURL", mock.Anything).Maybe().
-		Return(func(state string) string {
-			return "https://provider.com/o/oauth2/auth?" +
-				"code=someCode" +
-				"&state=" + state
-		})
+	// Инициализация адаптеров
+	suite.initAdapters()
+}
+
+// initAdapters настраивает моки адаптеров
+func (suite *SuiteWithMocks) initAdapters() {
+	suite.Adapters.Oauth = mockOauth.NewProvider(suite.T())
+	suite.Adapters.Oauth.EXPECT().Name().Maybe().Return("mock")
+	suite.Adapters.Oauth.EXPECT().Exchange(mock.Anything).Maybe().Return(func(code string) (userr.OpenAuthToken, error) {
+		token, ok := suite.MockOauthTokens[code]
+		if !ok {
+			return userr.OpenAuthToken{}, errors.New("token not found")
+		}
+		return token, nil
+	})
+	suite.Adapters.Oauth.EXPECT().User(mock.Anything).Maybe().Return(func(token userr.OpenAuthToken) (userr.OpenAuthUser, error) {
+		user, ok := suite.MockOauthUsers[token]
+		if !ok {
+			return userr.OpenAuthUser{}, errors.New("user not found")
+		}
+		return user, nil
+	})
+	suite.Adapters.Oauth.EXPECT().AuthorizationURL(mock.Anything).Maybe().Return(func(state string) string {
+		return "https://provider.com/o/oauth2/auth?" +
+			"code=someCode" +
+			"&state=" + state
+	})
 
 	// Тестовые пользователи и токены
 	suite.MockOauthUsers = suite.GenerateMockUsers()
@@ -141,19 +97,8 @@ func (suite *Suite) SetupTest() {
 	}
 }
 
-// TearDownSubTest выполняется после каждого подтеста, связанного с suite
-func (suite *Suite) TearDownSubTest() {
-	err := suite.factory.Cleanup()
-	suite.Require().NoError(err)
-}
-
-// TearDownTest выполняется после каждого подтеста, связанного с suite
-func (suite *Suite) TearDownTest() {
-	suite.factoryCloser()
-}
-
 // EqualSessions сравнивает две сессии
-func (suite *Suite) EqualSessions(s1, s2 sessionn.Session) {
+func (suite *SuiteWithMocks) EqualSessions(s1, s2 sessionn.Session) {
 	suite.Equal(s1.ID, s2.ID)
 	suite.Equal(s1.UserID, s2.UserID)
 	suite.Equal(s1.Name, s2.Name)
@@ -164,16 +109,8 @@ func (suite *Suite) EqualSessions(s1, s2 sessionn.Session) {
 	suite.True(s1.RefreshToken.Expiry.Equal(s2.RefreshToken.Expiry))
 }
 
-// UpsertChat сохраняет чат в репозиторий, в случае ошибки завершит тест
-func (suite *Suite) UpsertChat(chat chatt.Chat) chatt.Chat {
-	err := suite.RR.Chats.Upsert(chat)
-	suite.Require().NoError(err)
-
-	return chat
-}
-
 // RndChat создает случайный чат
-func (suite *Suite) RndChat() chatt.Chat {
+func (suite *SuiteWithMocks) RndChat() chatt.Chat {
 	chat, err := chatt.NewChat(gofakeit.Noun(), uuid.New(), nil)
 	suite.Require().NoError(err)
 
@@ -181,14 +118,14 @@ func (suite *Suite) RndChat() chatt.Chat {
 }
 
 // NewParticipant создает случайного участника
-func (suite *Suite) NewParticipant(userID uuid.UUID) chatt.Participant {
+func (suite *SuiteWithMocks) NewParticipant(userID uuid.UUID) chatt.Participant {
 	p, err := chatt.NewParticipant(userID)
 	suite.Require().NoError(err)
 	return p
 }
 
 // AddRndParticipant добавляет случайного участника в чат
-func (suite *Suite) AddRndParticipant(chat *chatt.Chat) chatt.Participant {
+func (suite *SuiteWithMocks) AddRndParticipant(chat *chatt.Chat) chatt.Participant {
 	p, err := chatt.NewParticipant(uuid.New())
 	suite.Require().NoError(err)
 	suite.Require().NoError(chat.AddParticipant(p, nil))
@@ -197,24 +134,24 @@ func (suite *Suite) AddRndParticipant(chat *chatt.Chat) chatt.Participant {
 }
 
 // AddParticipant добавляет участника в чат
-func (suite *Suite) AddParticipant(chat *chatt.Chat, p chatt.Participant) {
+func (suite *SuiteWithMocks) AddParticipant(chat *chatt.Chat, p chatt.Participant) {
 	suite.Require().NoError(chat.AddParticipant(p, nil))
 }
 
 // NewInvitation создает новое приглашение
-func (suite *Suite) NewInvitation(subjectID, recipientID uuid.UUID) chatt.Invitation {
+func (suite *SuiteWithMocks) NewInvitation(subjectID, recipientID uuid.UUID) chatt.Invitation {
 	i, err := chatt.NewInvitation(subjectID, recipientID)
 	suite.Require().NoError(err)
 	return i
 }
 
 // AddInvitation добавляет приглашение в чат
-func (suite *Suite) AddInvitation(chat *chatt.Chat, i chatt.Invitation) {
+func (suite *SuiteWithMocks) AddInvitation(chat *chatt.Chat, i chatt.Invitation) {
 	suite.Require().NoError(chat.AddInvitation(i, nil))
 }
 
 // RandomString генерирует случайную строку
-func RandomString(n int) string {
+func RandomString2(n int) string {
 	const letters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
 	b := make([]byte, n)
 	for i := range b {
@@ -224,7 +161,7 @@ func RandomString(n int) string {
 }
 
 // RandomOauthToken генерирует случайный OauthToken
-func (suite *Suite) RandomOauthToken() userr.OpenAuthToken {
+func (suite *SuiteWithMocks) RandomOauthToken() userr.OpenAuthToken {
 	t, err := userr.NewOpenAuthToken(
 		RandomString(32), // AccessToken
 		"Bearer",         // TokenType
@@ -238,7 +175,7 @@ func (suite *Suite) RandomOauthToken() userr.OpenAuthToken {
 }
 
 // Генерация случайного OauthUser
-func (suite *Suite) RandomOauthUser() userr.OpenAuthUser {
+func (suite *SuiteWithMocks) RandomOauthUser() userr.OpenAuthUser {
 	u, err := userr.NewOpenAuthUser(
 		RandomString(21),                                      // ID
 		suite.Adapters.Oauth.Name(),                           // Provider
@@ -253,7 +190,7 @@ func (suite *Suite) RandomOauthUser() userr.OpenAuthUser {
 }
 
 // GenerateMockUsers генерирует случайных oauth-пользователей
-func (suite *Suite) GenerateMockUsers() map[userr.OpenAuthToken]userr.OpenAuthUser {
+func (suite *SuiteWithMocks) GenerateMockUsers() map[userr.OpenAuthToken]userr.OpenAuthUser {
 	tokenToUser := make(map[userr.OpenAuthToken]userr.OpenAuthUser)
 
 	for i := 0; i < 10; i++ {
@@ -266,7 +203,7 @@ func (suite *Suite) GenerateMockUsers() map[userr.OpenAuthToken]userr.OpenAuthUs
 }
 
 // NewRndUserWithBasicAuth создает случайного пользователя с базовой аутентификацией
-func (suite *Suite) NewRndUserWithBasicAuth() userr.User {
+func (suite *SuiteWithMocks) NewRndUserWithBasicAuth() userr.User {
 	user, err := userr.NewUser(gofakeit.Name(), gofakeit.Username())
 	suite.Require().NoError(err)
 	ba, err := userr.NewBasicAuth(gofakeit.Username()+"four", common.RndPassword())
@@ -277,7 +214,7 @@ func (suite *Suite) NewRndUserWithBasicAuth() userr.User {
 }
 
 // HasElementOfType возвращает true, если в срезе есть элемент заданного типа
-func HasElementOfType[T any](e []any) bool {
+func HasElementOfType2[T any](e []any) bool {
 	for _, e := range e {
 		if _, ok := e.(T); ok {
 			return true
@@ -286,9 +223,19 @@ func HasElementOfType[T any](e []any) bool {
 	return false
 }
 
-func (suite *Suite) AssertHasEventType(ee []events.Event, eventType string) {
+func (suite *SuiteWithMocks) AssertHasEventType(ee []events.Event, eventType string) {
 	suite.T().Helper()
 	suite.True(slices.ContainsFunc(ee, func(e events.Event) bool {
 		return e.Type == eventType
 	}))
+}
+
+// RandomString генерирует случайную строку
+func RandomString(n int) string {
+	const letters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+	b := make([]byte, n)
+	for i := range b {
+		b[i] = letters[rand.Intn(len(letters))]
+	}
+	return string(b)
 }
